test: single test_drug_sensitivity_dataset.py.
